# server/swagger/swagger.yaml
definitions:
...
  api.Meta:
    properties:
      error:
        type: string
      msg:
        description: Status int64  `json:"status"`
        type: string
    type: object
...
  api.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/api.Meta'
    type: object
  api.SubmitTaskReq:
    properties:
      auditorId:
        items:
          type: integer
        type: array
      tid:
        type: integer
      uid:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/ops/approveTask:
    put:
      description: 传入工单的ID
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 用户审批工单
      tags:
      - Ops相关
  /api/v1/ops/delete:
    delete:
      description: 传入工单的ID
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 工单删除
      tags:
      - Ops相关
  /api/v1/ops/execTask:
    post:
      description: 返回执行结果
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 工单操作执行
      tags:
      - Ops相关
  /api/v1/ops/getExecParam:
    get:
      description: 返回sftp和ssh的执行参数
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 提取执行参数
      tags:
      - Ops相关
  /api/v1/ops/getTask:
    get:
      description: 传入查询所需参数,输了ID就不用name和页码
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: page_size
        type: integer
      - in: query
        name: task_name
        type: string
      - in: query
        name: tid
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 查看任务工单
      tags:
      - Ops相关
  /api/v1/ops/submitTask:
    post:
      description: 传入模板id，返回ssh执行所需参数并自动写入任务工单库
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: '注意Auditor参数: 最先审批的放第一个,因为接入后从第一个到最后一个依次发送信息审批'
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.SubmitTaskReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 提交执行工单
      tags:
      - Ops相关
...
swagger: "2.0"